+++ Frends.AmazonSQS.Receive/Frends.AmazonSQS.Receive/Definitions/Options.cs
@@ -8,6 +8,22 @@ namespace Frends.AmazonSQS.Receive.Definitions;
 public class Options
 {
     /// <summary>
+    /// Gets or sets a value indicating whether an error should stop the Task and throw an exception.
+    /// If set to true, an exception will be thrown when an error occurs. If set to false, Task will try to continue and the error message will be added into Result.Error and Result.Success will be set to false.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Gets or sets a custom error message to be used when an error occurs and ThrowErrorOnFailure is set to false.
+    /// If empty, the original error message will be used.
+    /// </summary>
+    /// <example>Failed to receive message from AmazonSQS.</example>
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; }
+
+    /// <summary>
     /// Delete message(s) from queue after receiving it 
     /// </summary>
     /// <example>true</example>