+++ Frends.AmazonSQS.Receive/Frends.AmazonSQS.Receive/Definitions/Result.cs
@@ -1,7 +1,5 @@
-﻿using Amazon.Runtime;
-using Amazon.SQS.Model;
+﻿using Amazon.SQS.Model;
 using System.Collections.Generic;
-using System.Net;
 
 namespace Frends.AmazonSQS.Receive.Definitions;
 
@@ -11,6 +9,12 @@ namespace Frends.AmazonSQS.Receive.Definitions;
 public class Result
 {
     /// <summary>
+    /// Gets a value indicating whether the Task was executed successfully.
+    /// </summary>
+    /// <example>true</example>
+    public bool Success { get; private set; }
+
+    /// <summary>
     /// Length of the received content.
     /// </summary>
     /// <example></example>
@@ -19,26 +23,41 @@ public class Result
     /// <summary>
     /// Status code of the operation.
     /// </summary>
-    /// <example></example>
-    public HttpStatusCode StatusCode { get; private set; }
+    /// <example>200</example>
+    public int StatusCode { get; private set; }
 
     /// <summary>
     /// List of received messages.
     /// </summary>
     /// <example></example>
-    public List<Message> Messages { get; private set; }
+    public List<ReceivedMessage> Messages { get; private set; }
 
     /// <summary>
     /// Information about the request.
     /// </summary>
     /// <example></example>
-    public ResponseMetadata ResponseMetadata { get; private set; }
+    public SqsResponseMetadata ResponseMetadata { get; private set; }
+
+    /// <summary>
+    /// Error information.
+    /// This value is generated when an exception occurs and Options.ThrowErrorOnFailure = false.
+    /// </summary>
+    /// <example>Error occurred...</example>
+    public Error Error { get; private set; }
 
     internal Result(ReceiveMessageResponse response)
     {
+        Success = true;
         ContentLength = response.ContentLength;
-        StatusCode = response.HttpStatusCode;
-        Messages = response.Messages;
-        ResponseMetadata = response.ResponseMetadata;
+        StatusCode = (int)response.HttpStatusCode;
+        Messages = response.Messages.ConvertAll(m => new ReceivedMessage(m));
+        ResponseMetadata = new SqsResponseMetadata(response.ResponseMetadata);
+        Error = null;
+    }
+
+    internal Result(Error error)
+    {
+        Success = false;
+        Error = error;
     }
 }